requirements.txt
requests
streamlit
soundfile
suno
ollama
torch
pandas
poe_api_wrapper
asyncio
assemblyai
python-doctr[torch]
opencv-python
numpy
SunoAI